refactor(dashboard page_users.py) split code file
make a directory for the user page instead, so that the code files
get easier to understand

improve logging by using a named logger and exception logging
