+++ odtp/dashboard/main.py
@@ -7,7 +7,7 @@ from odtp.dashboard.pages.page_components import content as components_page
 from odtp.dashboard.pages.page_digital_twins import content as digital_twins_page
 from odtp.dashboard.pages.page_executions import content as executions_page
 from odtp.dashboard.pages.page_run import content as run_page
-from odtp.dashboard.pages.page_user import content as user_page
+from odtp.dashboard.page_users.main import content as user_page
 
 
 @ui.page("/")
